chore: remove legacy junit fixtures and runner
